% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{compute.xty}
\alias{compute.xty}
\title{Compute X'y Vector}
\usage{
compute.xty(x, y, robust = 0, k_value = 1.5)
}
\arguments{
\item{x}{Matrix, input data matrix}

\item{y}{Vector, response vector}

\item{robust}{Integer, 0 for classical estimate, 1 for Huber robust estimate}

\item{k_value}{Numeric, tuning parameter for Huber function}
}
\value{
Covariance vector
}
\description{
Compute X'y Vector
}
