% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{fit_standardize_x}
\alias{fit_standardize_x}
\title{Apply Standardization to New Data}
\usage{
fit_standardize_x(x, x.mean, x.sd)
}
\arguments{
\item{x}{Matrix, input data matrix to be standardized}

\item{x.mean}{Vector, column means from training data}

\item{x.sd}{Vector, column standard deviations from training data}
}
\value{
Standardized matrix
}
\description{
Apply Standardization to New Data
}
