% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{generate.cov}
\alias{generate.cov}
\title{Generate Covariance Matrix}
\usage{
generate.cov(p, example)
}
\arguments{
\item{p}{Integer, dimension of the covariance matrix}

\item{example}{Integer, type of covariance structure (1=AR(1), 2=Block diagonal, 3=Kronecker product)}
}
\value{
A p x p covariance matrix
}
\description{
Generate Covariance Matrix
}
