% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_block_indices}
\alias{get_block_indices}
\title{Get Block Indices}
\usage{
get_block_indices(pp)
}
\arguments{
\item{pp}{Vector, block sizes}
}
\value{
List with start and end indices for each block
}
\description{
Get Block Indices
}
