% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{im_mean}
\alias{im_mean}
\title{Mean Imputation}
\usage{
im_mean(X)
}
\arguments{
\item{X}{Matrix with missing values}
}
\value{
Matrix with missing values imputed by column means
}
\description{
Mean Imputation
}
