% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{im_svd}
\alias{im_svd}
\title{SVD Imputation}
\usage{
im_svd(X)
}
\arguments{
\item{X}{Matrix with missing values}
}
\value{
Matrix with missing values imputed using SVD
}
\description{
SVD Imputation
}
