% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{standardize_x}
\alias{standardize_x}
\title{Standardize Matrix}
\usage{
standardize_x(x, robust = 0, k.value = 1.5)
}
\arguments{
\item{x}{Matrix, input data matrix}

\item{robust}{Integer, 0 for classical estimate, 1 for Huber robust estimate}

\item{k.value}{Numeric, tuning parameter for robust estimation}
}
\value{
List with standardized matrix and scaling parameters
}
\description{
Standardize Matrix
}
