% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_block_indices}
\alias{get_block_indices}
\title{Get Block Indices}
\usage{
get_block_indices(pp)
}
\arguments{
\item{pp}{Vector, block sizes}
}
\value{
List with start and end indices for each block
}
\description{
Get Block Indices
}
\examples{
# Define block sizes
pp <- c(10, 15, 20)
indices <- get_block_indices(pp)
print(indices)
# Shows: $starts = c(1, 11, 26) and $ends = c(10, 25, 45)

# For two blocks
pp2 <- c(25, 25)
indices2 <- get_block_indices(pp2)
print(indices2)
}
